% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraph.R
\name{makeGraph}
\alias{makeGraph}
\title{Create a connected maze graph that represents a rectangular grid.
Nodes represent cells and edges betweem nodes represent wall sites.
All walls are initially on.}
\usage{
makeGraph(nrows = 0, ncols = 0)
}
\arguments{
\item{nrows}{maze hight (number of rows); default value set to 0.}

\item{ncols}{maze width (number of columns); default value set to 0.}
}
\value{
This function creates and returns a maze graph that represents a rectangular grid
 that matches user specified maze dimensions. Nodes in the graph will be named in the
 format Aij, where i corresponds to a row number  and j corresponds to a column number.
}
\description{
Create a connected maze graph that represents a rectangular grid.
Nodes represent cells and edges betweem nodes represent wall sites.
All walls are initially on.
}
\examples{
maze1 <- makeGraph(10, 10)
maze1 <- makeMaze_dfs(maze1)
plotMaze(maze1, 10, 10)

}

