% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImperfect.R
\name{makeImperfect}
\alias{makeImperfect}
\title{Given a perfect maze graph, create an imperfect maze graph.}
\usage{
makeImperfect(gD = NA, ptc = 20, inShiny = FALSE)
}
\arguments{
\item{gD}{an existing maze graph object.}

\item{ptc}{percentage of walls to randomly add or remove; default value is 10(percent).}

\item{inShiny}{a flag that marks whether the function is called from a shiny app or console.}
}
\value{
A maze graph object containing an imperfect maze.
}
\description{
Given a perfect maze graph, create an imperfect maze graph.
}
\examples{
maze1 <- makeGraph(10, 10)
maze1 <- makeMaze_dfs(maze1)
plotMaze(maze1, 10, 10)
maze1 <- makeImperfect(maze1)
plotMaze(maze1, 10, 10)

}

