% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaze.R
\name{plotMaze}
\alias{plotMaze}
\title{Given a maze graph, plot a maze.}
\usage{
plotMaze(gD = NA, nrows = 0, ncols = 0, inShiny = FALSE)
}
\arguments{
\item{gD}{an existing maze graph object.}

\item{nrows}{maze hight (number of rows); default value set to 0.}

\item{ncols}{maze width (number of columns); default value set to 0.}

\item{inShiny}{a flag that marks whether the function is called from a shiny app or console}
}
\value{
This function uses ggplot to plot a maze. Currently, maze enterance and exit
 points (cells/nodes) are fixed.
}
\description{
Given a maze graph, plot a maze.
}
\examples{
maze1 <- makeGraph(10, 10)
maze1 <- makeMaze_dfs(maze1)
plotMaze(maze1, 10, 10)

}

