% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMazeSolution.R
\name{plotMazeSolution}
\alias{plotMazeSolution}
\title{Plot maze solution (path)}
\usage{
plotMazeSolution(gD = NA, nrows = 0, ncols = 0, inShiny = FALSE)
}
\arguments{
\item{gD}{an existing maze graph object.}

\item{nrows}{maze hight (number of rows); default value set to 0.}

\item{ncols}{maze width (number of columns); default value set to 0.}

\item{inShiny}{a flag that marks whether the function is called from a shiny app or console.}
}
\value{
The maze solution is found as the shortest path between the start and end maze cells (currently, start and
 end points of the maze are fixed). The function uses ggplot to plot a maze.
}
\description{
Plot maze solution (path)
}
\examples{
maze1 <- makeGraph(10, 10)
maze1 <- makeMaze_dfs(maze1)
plotMaze(maze1, 10, 10)
plotMazeSolution(maze1, 10, 10)

}

