% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{A function to start a Shiny app that comes with the package.
Code from: http://www.r-bloggers.com/supplementing-your-r-package-with-a-shiny-app-2/}
\usage{
runExample()
}
\description{
A function to start a Shiny app that comes with the package.
Code from: http://www.r-bloggers.com/supplementing-your-r-package-with-a-shiny-app-2/
}
\examples{
Rmaze::runExample()
}

